refactor[#523] : entities/follow queryKey constants -> api 로 이동
