+++ src/entities/follow/api/getFollowerList.ts
@@ -1,8 +1,9 @@
 import { skipToken, useInfiniteQuery } from "@tanstack/react-query";
 import { apiClient } from "@/shared/lib";
 import { useAuthStore } from "@/shared/store";
-import { FOLLOW_END_POINT, followQueryKey } from "../constants";
+import { FOLLOW_END_POINT } from "../constants";
 import type { FollowUserInfo } from "../types/server";
+import { followQueryKey } from "./queryKey";
 
 // TODO 리팩토링 시 해당 타입 정의 위치 의논
 interface PageAbleInformation {
