+++ src/entities/follow/constants/index.ts
@@ -1,2 +1,1 @@
 export * from "./endpoint";
-export * from "./queryKey";
